requirements.txt
Django==3.1.3
django-environ==0.4.5
qrcode==6.1
pillow==8.0.1
